refactor: extract lancedb-dump-service to packages/application
